Rename agents/v1 directory to agents/v010 for versioning clarity
- Renamed all module files from agents/v1 to agents/v010
- Updated imports and documentation references to reflect new path
- Updated refactoring analysis document with correct import path
- Fixed file permissions on prompt_formatter.py

This change distinguishes the v010 (0.1.0) implementation from future
v1 (1.0) implementations of the agent framework, providing clearer
version semantics.
