+++ agents/v010/agent_base.py
@@ -11,7 +11,7 @@ import cosa.memory.solution_snapshot as ss
 
 from cosa.agents.llm_v0 import Llm_v0
 from cosa.agents.raw_output_formatter import RawOutputFormatter
-from cosa.agents.v1.runnable_code import RunnableCode
+from cosa.agents.v010.runnable_code import RunnableCode
 from cosa.app.configuration_manager import ConfigurationManager
 from cosa.memory.solution_snapshot import SolutionSnapshot
 from cosa.agents.two_word_id_generator import TwoWordIdGenerator
