+++ agents/v010/date_and_time_agent.py
@@ -1,4 +1,4 @@
-from cosa.agents.v1.agent_base import AgentBase
+from cosa.agents.v010.agent_base import AgentBase
 
 class DateAndTimeAgent( AgentBase ):
